refactor: 불필요한 useState 정리
