+++ fe/src/components/Cart/CartItem.module.css
@@ -67,7 +67,7 @@
   padding: 0 3px;
   margin-top: 5px;
   font-size: 14px;
-  color: #444444;
+  color: #303345;
   text-align: center;
   word-break: keep-all;
 }
