+++ fe/src/components/Cart/CartLists.module.css
@@ -1,5 +1,5 @@
 .cartLists {
-  width: 435px;
+  width: 425px;
   height: 173px;
   background-color: #30334432;
   border-radius: 17px;
